+++ NEWS.md
@@ -14,6 +14,7 @@ Bug fixes:
 objects to avoid counts becoming inaccurate ([#332](https://github.com/timoast/signac/issues/332))
 * Prevent dropping features when creating a merged ChromatinAssay ([#340](https://github.com/timoast/signac/pull/340))
 * Fix compilation error when using g++ version <5 ([#326](https://github.com/timoast/signac/issues/326))
+* Retain motif positions during subset ([#364](https://github.com/timoast/signac/issues/364))
 
 Other changes: